Fix all AMP dtype mismatches in model.py
Checked every torch.zeros, indexed assignment, and UNK embed usage.
Fixed three spots:
1. build_arg_embeddings: arg_emb dtype matches symbol_embeds
2. build_arg_embeddings: unk_sym/var_embed cast to match target dtype
3. _pad_symbol_embeds: padded tensor dtype matches embeds

Verified: loss function (F.cross_entropy) handles mixed precision
internally. Encoder inputs are fp32 (raw data), cast by autocast
in linear layers. Generate path doesn't use autocast.

+++ conjecture_gen/model.py
@@ -446,7 +446,8 @@ class PointerTreeDecoder(nn.Module):
         if max_n == 0:
             max_n = 1
 
-        padded = torch.zeros(batch_size, max_n, self.hidden_dim, device=device)
+        padded = torch.zeros(batch_size, max_n, self.hidden_dim,
+                             device=device, dtype=embeds.dtype)
         mask = torch.zeros(batch_size, max_n, dtype=torch.bool, device=device)
 
         sorted_idx = torch.argsort(batch_assign, stable=True)
